add grub; do some refactoring

+++ pi2clone.sh
@@ -32,6 +32,9 @@ Where:
     -i  Interactive, showing progress bars
 "
 
+#DEBUG ONLY
+#printarr() { declare -n __p="$1"; for k in "${!__p[@]}"; do printf "%s=%s\n" "$k" "${__p[$k]}" ; done ;  }
+
 usage() {
     printf "%s\n" "$USAGE"
     exit 1
@@ -96,10 +99,10 @@ to_file() {
                 mount "$sdev" -t "${filesystems[$sdev]}" "/mnt/$sdev"
             fi
 
-            cmd="tar --directory=/mnt/$tdev --exclude=proc/* --exclude=dev/* --exclude=sys/* --atime-preserve --numeric-owner --xattrs"
+            cmd="tar --directory=/mnt/$tdev --exclude=/proc/* --exclude=/dev/* --exclude=/sys/* --atime-preserve --numeric-owner --xattrs"
 
             if $INTERACTIVE; then 
-                local size=$(du --bytes --exclude=proc/* --exclude=sys/* -s /mnt/$tdev | tr -s '\t' ' ' | cut -d ' ' -f 1)
+                local size=$(du --bytes --exclude=/proc/* --exclude=/sys/* -s /mnt/$tdev | tr -s '\t' ' ' | cut -d ' ' -f 1)
                 cmd="$cmd -Scpf - . | pv --rate --timer --eta -pe -s $size > ${i}.${sid:-NOUUID}.${spid:-NOPUUID}.${fs}.${type}.${sdev//\//_}" 
             else
                 cmd="$cmd -Scpf ${i}.${sid:-NOUUID}.${spid:-NOPUUID}.${fs}.${type}.${sdev//\//_} ."  
@@ -258,41 +261,67 @@ from_file() {
 clone() {
     local src="$1" dest="$2"
     declare -A filesystems mounts partuuids uuids types
-    declare -A src_lfs dest_lfs dests src2dest psrc2pdest
+    declare -A src_lfs dest_lfs dests src2dest psrc2pdest nsrc2ndest
 
     local spuuids=() suuids=() 
     local dpuuids=() duuids=()
-    local sfs=() lmbrs=() srcs=() ldests=()
+    local sfs=() lmbrs=() srcs=() ldests=() lsrcs=()
     local vg_src_name vg_src_name_clone
+    local has_grub=false
 
-    _set_uuids() {
-        local src_dest=$1
-        local mkfs=$2
-        [[ $src_dest == dest ]] && dpuuids=() duuids=()
-        [[ $src_dest == src ]] && spuuids=() suuids=()
-
+    _set_dest_uuids() {
+        dpuuids=() duuids=() dnames=()
         while read -r e; do
             read -r kdev dev fstype uuid puuid type parttype mnt<<< "$e"
             eval "$kdev" "$dev" "$fstype" "$uuid" "$puuid" "$type" "$parttype" "$mountpoint"
-            [[ $FSTYPE == swap ]] && continue
-            if [[ $src_dest == dest ]]; then
-                [[ -n $UUID ]] && dests[$UUID]=$NAME
-                dpuuids+=($PARTUUID)
-                duuids+=($UUID)
+            [[ $PARTTYPE == 0x5 || $TYPE == disk ]] && continue
+            [[ -n $UUID ]] && dests[$UUID]=$NAME
+            dpuuids+=($PARTUUID)
+            duuids+=($UUID)
+            dnames+=($NAME)
+        done < <( lsblk -Ppo KNAME,NAME,FSTYPE,UUID,PARTUUID,TYPE,PARTTYPE,MOUNTPOINT "$dest" )
+    }
+
+    _set_src_uuids() {
+        spuuids=() suuids=() snames=()
+        while read -r e; do
+            read -r kdev dev fstype uuid puuid type parttype mnt<<< "$e"
+            eval "$kdev" "$dev" "$fstype" "$uuid" "$puuid" "$type" "$parttype" "$mountpoint"
+            [[ $PARTTYPE == 0x5 || $TYPE == disk ]] && continue
+            [[ $FSTYPE == LVM2_member ]] && lmbrs[${NAME: -1}]="$UUID"
+            [[ $TYPE == part && $FSTYPE != LVM2_member ]] && sfs[${NAME: -1}]=$FSTYPE
+            spuuids+=($PARTUUID)
+            suuids+=($UUID)
+            snames+=($NAME)
+        done < <( lsblk -Ppo KNAME,NAME,FSTYPE,UUID,PARTUUID,TYPE,PARTTYPE,MOUNTPOINT "$src" )
+    }
+
+    _init_srcs() {
+        while read -r e; do
+            read -r kdev dev fstype uuid puuid type parttype mountpoint<<< "$e"
+            eval "$kdev" "$dev" "$fstype" "$uuid" "$puuid" "$type" "$parttype" "$mountpoint"
+            [[ $PARTTYPE == 0x5 || $FSTYPE == LVM2_member ]] && continue #82 extended, 5 swap
+            [[ $TYPE == lvm ]] && lsrcs+=($NAME) && islvm=true
+            [[ $TYPE == part ]] && srcs+=($NAME)
+            filesystems[$NAME]="$FSTYPE"
+            [[ -n $MOUNTPOINT ]] && mounts[$MOUNTPOINT]="$UUID"
+            partuuids[$NAME]="$PARTUUID"
+            uuids[$NAME]="$UUID"
+            types[$NAME]="$TYPE"
+        done < <(lsblk -Ppo KNAME,NAME,FSTYPE,UUID,PARTUUID,TYPE,PARTTYPE,MOUNTPOINT "$src")
+    }
+
+    _mkfs() {
+        while read -r e; do
+            read -r kdev dev fstype uuid puuid type parttype mnt<<< "$e"
+            eval "$kdev" "$dev" "$fstype" "$uuid" "$puuid" "$type" "$parttype" "$mountpoint"
+            if [[ ${sfs[${NAME: -1}]} == swap ]]; then
+                mkswap "$NAME"
+            else
+                [[ ${sfs[${NAME: -1}]} ]] && mkfs -t "${sfs[${NAME: -1}]}" "$NAME"
+                [[ ${lmbrs[${NAME: -1}]} ]] && pvcreate "$NAME" 2>/dev/null
             fi
-            if [[ $src_dest == src ]]; then
-                [[ $FSTYPE == LVM2_member ]] && lmbrs[${NAME: -1}]="$UUID"
-                [[ $TYPE == part && $FSTYPE != LVM2_member ]] && sfs[${NAME: -1}]=$FSTYPE
-                spuuids+=($PARTUUID)
-                suuids+=($UUID)
-            fi
-            if [[ -n $mkfs ]]; then
-                if [[ $TYPE == part ]]; then
-                    [[ ${sfs[${NAME: -1}]} ]] && mkfs -t "${sfs[${NAME: -1}]}" "$NAME"
-                    [[ ${lmbrs[${NAME: -1}]} ]] && pvcreate "$NAME" 2>/dev/null
-                fi
-            fi
-        done < <( eval lsblk -Ppo KNAME,NAME,FSTYPE,UUID,PARTUUID,TYPE,PARTTYPE,MOUNTPOINT "\$$src_dest" )
+        done < <( lsblk -Ppo KNAME,NAME,FSTYPE,UUID,PARTUUID,TYPE,PARTTYPE,MOUNTPOINT "$dest" )
     }
 
     _lvm_setup() {
@@ -305,28 +334,35 @@ clone() {
         done < <( pvs --noheadings -o pv_name,vg_name )
 
         while read -r e; do
-            read -r lv_name vg_name lv_size vg_free<<< "$e"
-            lvcreate -L "$lv_size" -n "$lv_name" "$vg_src_name_clone"
-        done < <( lvs --noheadings --units m --nosuffix -o lv_name,vg_name,lv_size,vg_free )
+            read -r lv_name vg_name lv_size vg_size vg_free<<< "$e"
+            lvcreate --yes -l$(echo "$lv_size * 100 / $vg_size" | bc)%VG -n "$lv_name" "$vg_src_name_clone"
+        done < <( lvs --noheadings --units m --nosuffix -o lv_name,vg_name,lv_size,vg_size,vg_free )
 
         for d in $SRC $DEST; do
             while read -r e; do
                 read -r kdev dev fstype type<<< "$e"
                 eval "$kdev" "$dev" "$fstype" "$type"
-                [[ $TYPE == 'lvm' && $d == "$SRC" ]] && src_lfs[${NAME##*$vg_src_name-}]=$FSTYPE 
-                [[ $TYPE == 'lvm' && $d == "$DEST" ]] && mkfs -t "${src_lfs[${NAME##*$vg_src_name_clone-}]}" "$NAME";
+                [[ $TYPE == 'lvm' && $d == "$SRC" ]] && src_lfs[${NAME##*-}]=$FSTYPE 
+                if [[ $TYPE == 'lvm' && $d == "$DEST" ]]; then
+                    { [[ "${src_lfs[${NAME##*-}]}" == swap ]] && mkswap "$NAME"; } || mkfs -t "${src_lfs[${NAME##*-}]}" "$NAME";
+                fi
             done < <( eval lsblk -Ppo KNAME,NAME,FSTYPE,TYPE "$d" )
         done
     }
 
     _prepare_disk() {
-        local vg_src_name=$(pvs --noheadings -o pv_name,vg_name | grep "$src" | xargs | cut -d ' ' -f2)
-        local vg_src_name_clone="${vg_src_name}_clone"
+        if hash lvm 2>/dev/null; then
+            local vg_src_name=$(pvs --noheadings -o pv_name,vg_name | grep "$src" | xargs | cut -d ' ' -f2)
+            local vg_src_name_clone="${vg_src_name}_clone"
 
-        vgchange -an "$vg_src_name_clone" 2>/dev/null
-        vgremove -f "$vg_src_name_clone" 2>/dev/null
+            vgchange -an "$vg_src_name_clone" 2>/dev/null
+            vgremove -f "$vg_src_name_clone" 2>/dev/null
+        fi
 
         dd if=/dev/zero of="$dest" bs=512 count=100000
+        #For some reason sfdisk < 2.29 does not create PARTUUIDs when importing a partition table.
+        #But when we create a partition and afterward import a prviously dumped partition table, it works!
+        echo -e "n\np\n\n\n\nw\n" | fdisk "$dest"
     }
 
     _boot_setup() {
@@ -335,113 +371,130 @@ clone() {
 
         for k in "${!sd[@]}"; do
             for d in "${dests[@]}"; do
-                sed -i "s/$k/${sd[$k]}/" "/mnt/$d/cmdline.txt" "/mnt/$d/etc/fstab" 2>/dev/null
+                sed -i "s|$k|${sd[$k]}|" \
+                    "/mnt/$d/cmdline.txt" "/mnt/$d/etc/fstab" \
+                    "/mnt/$d/grub/grub.cfg" "/mnt/$d/boot/grub/grub.cfg" "/mnt/$d/etc/initramfs-tools/conf.d/resume" \
+                    2>/dev/null
             done
         done
     }
 
-    _prepare_disk
+    _create_rclocal() {
+        mv $1/etc/rc.local $1/etc/rc.local.bak 2>/dev/null
+        printf '%s' '#! /usr/bin/env bash
+        while read -r e; do
+            read -r dev type<<< "$e"
+            eval "$dev" "$type"
+            if [[ $TYPE == disk ]] && dd bs=512 count=1 if=/dev/sda 2>/dev/null | strings | grep -q GRUB; then 
+                update-initramfs -u -k all
+                grub-install $NAME
+                update-grub
+            fi
+        done < <(lsblk -Ppo NAME,TYPE)
+        rm /etc/rc.local
+        mv /etc/rc.local.bak /etc/rc.local 2>/dev/null
+        reboot' > $1/etc/rc.local
+        chmod +x $1/etc/rc.local
+    }
 
+    _grub_setup() {
+        local d=$1
+        local b=$2
+
+        mount $d /mnt/$d
+        [[ -n $b ]] && mount $b /mnt/$d/boot
+
+        for f in sys dev dev/pts proc; do 
+            mount --bind /$f /mnt/$d/$f;
+        done
+        grub-install --boot-directory=/mnt/$d/boot $dest
+        # chroot /mnt/$d update-grub
+        chroot /mnt/$d apt-get install -y binutils
+        _create_rclocal "/mnt/$d"
+        umount -R /mnt/$d
+    }
+
+    _prepare_disk
     sleep 3
 
     sfdisk --force "$dest" < <(sfdisk -d "$src")
-
     sleep 3 #IMPORTANT !!! So changes by sfdisk can settle. 
             #Otherwise resultes from lsblk might still show old values!
 
-    while read -r e; do
-        read -r kdev dev fstype uuid puuid type parttype mnt<<< "$e"
-        eval "$kdev" "$dev" "$fstype" "$uuid" "$puuid" "$type" "$parttype" "$mountpoint"
-        [[ $FSTYPE == swap || $FSTYPE == LVM2_member ]] && continue
-        [[ $TYPE == lvm ]] && lsrcs+=($NAME) && islvm=true
-        [[ $TYPE == part ]] && srcs+=($NAME)
-        filesystems[$NAME]="$FSTYPE"
-        mounts[$NAME]="$MOUNTPOINT"
-        partuuids[$NAME]="$PARTUUID"
-        uuids[$NAME]="$UUID"
-        types[$NAME]="$TYPE"
-    done < <(lsblk -Ppo KNAME,NAME,FSTYPE,UUID,PARTUUID,TYPE,PARTTYPE,MOUNTPOINT "$src")
-
-    #First collect what we have in our backup
-    _set_uuids "src"
-
-    #Then create the filesystems and PVs
-    _set_uuids "dest" "mkfs" 
+    _init_srcs         #First collect what we have in our backup
+    _set_src_uuids
+    _mkfs              #Then create the filesystems and PVs
     sleep 3
-
-    #Now collect what we have created
-    _set_uuids "dest"
 
     if [[ ${#lmbrs[@]} -gt 0 ]]; then 
         _lvm_setup
         sleep 3
-        #... and after that collect what we have created, again
-        _set_uuids "dest"
     fi
 
-    if [[ ${#suuids[@]} != "${#duuids[@]}" || ${#spuuids[@]} != "${#dpuuids[@]}" ]]; then
-        echo "Source and destination tables for UUIDs or PARTUUIDs did not macht. This should not happen!"
+    _set_dest_uuids     #Now collect what we have created
+
+    if [[ ${#suuids[@]} != "${#duuids[@]}" || ${#spuuids[@]} != "${#dpuuids[@]}" || ${#snames[@]} != "${#dnames[@]}" ]]; then
+        echo "Source and destination tables for UUIDs, PARTUUIDs or NAMES did not macht. This should not happen!"
         exit 1
     fi
 
     for ((i=0;i<${#suuids[@]};i++)); do src2dest[${suuids[$i]}]=${duuids[$i]}; done
     for ((i=0;i<${#spuuids[@]};i++)); do psrc2pdest[${spuuids[$i]}]=${dpuuids[$i]}; done
+    for ((i=0;i<${#snames[@]};i++)); do nsrc2ndest[${snames[$i]}]=${dnames[$i]}; done
 
-    for x in srcs lsrcs; do
-        eval declare -n s="$x"
+    for x in "${srcs[@]}" "${lsrcs[@]}"; do
+        local sdev=$x
+        local sid=${uuids[$sdev]}
+        local ddev=${dests[${src2dest[$sid]}]}
 
-        for ((i=0;i<${#s[@]};i++)); do
-            local sdev=${s[$i]}
-            local sid=${uuids[$sdev]}
-            local spid=${partuuids[$sdev]}
-            local fs=${filesystems[$sdev]}
-            local type=${types[$sdev]}
+        mkdir -p "/mnt/$ddev" "/mnt/$sdev"
 
-            local ddev=${dests[${src2dest[$sid]}]}
-            mkdir -p "/mnt/$ddev"
+        [[ ${filesystems[$sdev]} == swap ]] && continue
+        if [[ $x == lsrcs && ${#lmbrs[@]} -gt 0 && "${vg_free%%.*}" -ge "500" ]]; then
+            echo "USING snapshot"
+            sdev='snap4clone'
             mkdir -p "/mnt/$sdev"
+            lvcreate -l100%FREE -s -n snap4clone "${vg_src_name}/$lv_name"
+            sleep 3
+            mount "/dev/${vg_src_name}/$sdev" "/mnt/$sdev"
+        else
+            mount "$sdev" "/mnt/$sdev"
+        fi
 
-            local tdev=$sdev
-            if [[ $x == lsrcs && ${#lmbrs[@]} -gt 0 && "${vg_free%%.*}" -ge "500" ]]; then
-                local tdev='snap4clone'
-                mkdir -p "/mnt/$tdev"
-                lvcreate -l100%FREE -s -n snap4clone "${vg_src_name}/$lv_name"
-                sleep 3
-                mount "/dev/${vg_src_name}/$tdev" -t "${filesystems[$sdev]}" "/mnt/$tdev"
-            else
-                mount "$sdev" -t "${filesystems[$sdev]}" "/mnt/$sdev"
-            fi
+        mount "$ddev" "/mnt/$ddev"
 
-            mount "$ddev" -t "${filesystems[$sdev]}" "/mnt/$ddev"
+        echo "CLONING $sdev"
+        if $INTERACTIVE; then
+            local size=$( \
+                    rsync -aSXxH --stats --dry-run "/mnt/$sdev/" "/mnt/$ddev" \
+                | grep -oP 'Number of files: \d*(,\d*)*' \
+                | cut -d ':' -f2 \
+                | tr -d ' ' \
+                | sed -e 's/,//' \
+            )
+            rsync -vaSXxH "/mnt/$sdev/" "/mnt/$ddev" | pv -lep -s "$size" >/dev/null
+        else
+            rsync -aSXxH "/mnt/$sdev/" "/mnt/$ddev"
+        fi
+        
+        sleep 3
 
+        umount "/mnt/$sdev/" 2>/dev/null
+        lvremove -f "${vg_src_name}/$sdev" 2> /dev/null
 
-            echo "Cloning $sdev"
-                if $INTERACTIVE; then
-                    local size=$( \
-                          rsync -aSXxH --stats --dry-run "/mnt/$tdev/" "/mnt/$ddev" \
-                        | grep -oP 'Number of files: \d*(,\d*)*' \
-                        | cut -d ':' -f2 \
-                        | tr -d ' ' \
-                        | sed -e 's/,//' \
-                    )
-                    rsync -vaSXxH "/mnt/$tdev/" "/mnt/$ddev" | pv -lep -s "$size" >/dev/null
-                else
-                    rsync -aSXxH "/mnt/$tdev/" "/mnt/$ddev"
-                fi
-                
-                sleep 3
-                umount "/mnt/$tdev/" 2>/dev/null
-                lvremove -f "${vg_src_name}/$tdev" 2> /dev/null
-
-            sed -i "s/$vg_src_name/$vg_src_name_clone/" "/mnt/$ddev/cmdline.txt" "/mnt/$ddev/etc/fstab" 2>/dev/null
-        done
+        [[ -f /mnt/$ddev/grub/grub.cfg || -f /mnt/$ddev/grub.cfg ]] && has_grub=true
+        sed -i "s/$vg_src_name/$vg_src_name_clone/" "/mnt/$ddev/cmdline.txt" "/mnt/$ddev/etc/fstab" 2>/dev/null
 
         _boot_setup "src2dest"
         _boot_setup "psrc2pdest"
+        _boot_setup "nsrc2ndest"
 
-        for ((i=0;i<${#s[@]};i++)); do umount "/mnt/${s[$i]}" 2>/dev/null; done
-        for ((i=0;i<${#s[@]};i++)); do umount "/mnt/${dests[${src2dest[${uuids[${s[$i]}]}]}]}" 2>/dev/null; done
+        umount "/mnt/$sdev" "/mnt/$ddev" 2>/dev/null
     done
+
+    if $has_grub; then 
+        _grub_setup ${dests[${src2dest[${mounts[/]}]}]} ${dests[${src2dest[${mounts[/boot]}]}]}
+    fi
 }
 
 main() {
@@ -457,7 +510,7 @@ main() {
 
 trap cleanup INT
 
-for c in rsync tar flock; do
+for c in rsync tar flock bc; do
     hash $c 2>/dev/null || { echo >&2 "ERROR: $c missing."; abort='exit 1'; }
 done
